Move aFileWithSizeAndNameHasBeenCreatedLocally to BasicStructure
